Add option to convert include paths based on the conan configuration.
One needs to provide path to conan_paths variable, which loads and parses the output. This is XLAB internal specific and depends on the output of the conan_paths.cmake file, which should look like:
set(<LIBRARY>_PACKAGE_DIR "<path>")
Include paths are then searched for the <path> and replaced with <LIBRARY>_PACKAGE_DIR_<conanfile_hash> where <conanfile_hash> is hashed conanfile.py.